[DAQ-1756] NumberAndUnitsComposite: change getValue() & setValue() to use double
Conversion for data binding is done in NumberUnitsWidgetProperty

+++ uk.ac.gda.client/src/uk/ac/gda/client/NumberUnitsWidgetProperty.java
@@ -33,7 +33,7 @@ public class NumberUnitsWidgetProperty<Q extends Quantity> extends WidgetValuePr
 	@Override
 	protected void doSetValue(Object source, Object value) {
 		if (source instanceof NumberAndUnitsComposite && value instanceof Double) {
-			((NumberAndUnitsComposite<Q>) source).setValue(value);
+			((NumberAndUnitsComposite<Q>) source).setValue((double) value);
 		}
 	}
